ml-service/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
fastapi==0.135.3
h11==0.16.0
idna==3.11
numpy==2.4.4
pillow==12.2.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.22
requests==2.33.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
